+++ packages/stream_chat_flutter/lib/src/message_input.dart
@@ -9,13 +9,11 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:photo_manager/photo_manager.dart';
 import 'package:shimmer/shimmer.dart';
 import 'package:stream_chat_flutter/src/commands_overlay.dart';
 import 'package:stream_chat_flutter/src/emoji/emoji.dart';
 import 'package:stream_chat_flutter/src/emoji_overlay.dart';
 import 'package:stream_chat_flutter/src/extension.dart';
-import 'package:stream_chat_flutter/src/media_list_view.dart';
 import 'package:stream_chat_flutter/src/message_list_view.dart';
 import 'package:stream_chat_flutter/src/multi_overlay.dart';
 import 'package:stream_chat_flutter/src/quoted_message_widget.dart';
@@ -77,6 +75,20 @@ typedef UserMentionTileBuilder = Widget Function(
 typedef ActionButtonBuilder = Widget Function(
   BuildContext context,
   IconButton defaultActionButton,
+);
+
+/// Widget builder for widgets that require may required data from the
+/// [MessageInputController]
+typedef MessageRelatedBuilder = Widget Function(
+  BuildContext context,
+  MessageInputController messageInputController,
+);
+
+/// Widget builder for a custom attachment picker.
+typedef AttachmentsPickerBuilder = Widget Function(
+  BuildContext context,
+  MessageInputController messageInputController,
+  StreamAttachmentPicker defaultPicker,
 );
 
 /// Location for actions on the [MessageInput]
@@ -204,6 +216,8 @@ class MessageInput extends StatefulWidget {
     this.commandButtonBuilder,
     this.customOverlays = const [],
     this.mentionAllAppUsers = false,
+    this.attachmentsPickerBuilder,
+    this.sendButtonBuilder,
   })  : assert(
           initialMessage == null || editMessage == null,
           "Can't provide both `initialMessage` and `editMessage`",
@@ -322,6 +336,12 @@ class MessageInput extends StatefulWidget {
   /// Defaults to false.
   final bool mentionAllAppUsers;
 
+  /// Builds bottom sheet when attachment picker is opened.
+  final AttachmentsPickerBuilder? attachmentsPickerBuilder;
+
+  /// Builder for creating send button
+  final MessageRelatedBuilder? sendButtonBuilder;
+
   @override
   MessageInputState createState() => MessageInputState();
 
@@ -349,7 +369,6 @@ class MessageInputState extends State<MessageInput> {
   Command? _chosenCommand;
   bool _actionsShrunk = false;
   bool _openFilePickerSection = false;
-  int _filePickerIndex = 0;
 
   /// The editing controller passed to the input TextField
   late final MessageInputController messageInputController =
@@ -524,7 +543,7 @@ class MessageInputState extends State<MessageInput> {
               widget.actionsLocation == ActionsLocation.right)
             _buildExpandActionsButton(context),
           if (widget.sendButtonLocation == SendButtonLocation.outside)
-            _animateSendButton(context),
+            _buildSendButton(context),
         ],
       );
 
@@ -588,25 +607,18 @@ class MessageInputState extends State<MessageInput> {
         ],
       );
 
-  Widget _animateSendButton(BuildContext context) {
-    late Widget sendButton;
-    if (_timeOut > 0) {
-      sendButton = _CountdownButton(count: _timeOut);
-    } else if (!_messageIsPresent &&
-        messageInputController.attachments.isEmpty) {
-      sendButton = widget.idleSendButton ?? _buildIdleSendButton(context);
-    } else {
-      sendButton = widget.activeSendButton != null
-          ? InkWell(
-              onTap: sendMessage,
-              child: widget.activeSendButton,
-            )
-          : _buildSendButton(context);
+  Widget _buildSendButton(BuildContext context) {
+    if (widget.sendButtonBuilder != null) {
+      return widget.sendButtonBuilder!(context, messageInputController);
     }
 
-    return AnimatedSwitcher(
-      duration: _streamChatTheme.messageInputTheme.sendAnimationDuration!,
-      child: sendButton,
+    return StreamMessageSendButton(
+      onSendMessage: sendMessage,
+      timeOut: _timeOut,
+      isIdle: !_messageIsPresent && messageInputController.attachments.isEmpty,
+      isEditEnabled: widget.editMessage != null,
+      idleSendButton: widget.idleSendButton,
+      activeSendButton: widget.activeSendButton,
     );
   }
 
@@ -817,7 +829,7 @@ class MessageInputState extends State<MessageInput> {
               widget.actionsLocation == ActionsLocation.rightInside)
             _buildExpandActionsButton(context),
           if (widget.sendButtonLocation == SendButtonLocation.inside)
-            _animateSendButton(context),
+            _buildSendButton(context),
         ],
       ),
     ).merge(passedDecoration);
@@ -946,236 +958,33 @@ class MessageInputState extends State<MessageInput> {
   }
 
   Widget _buildFilePickerSection() {
-    final _attachmentContainsFile =
-        messageInputController.attachments.any((it) => it.type == 'file');
-
-    final attachmentLimitCrossed =
-        messageInputController.attachments.length >= widget.attachmentLimit;
-
-    Color _getIconColor(int index) {
-      final streamChatThemeData = _streamChatTheme;
-      switch (index) {
-        case 0:
-          return messageInputController.attachments.isEmpty
-              ? streamChatThemeData.colorTheme.accentPrimary
-              : (!_attachmentContainsFile
-                  ? streamChatThemeData.colorTheme.accentPrimary
-                  : streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.2));
-        case 1:
-          return _attachmentContainsFile
-              ? streamChatThemeData.colorTheme.accentPrimary
-              : (messageInputController.attachments.isEmpty
-                  ? streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.5)
-                  : streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.2));
-        case 2:
-          return attachmentLimitCrossed
-              ? streamChatThemeData.colorTheme.textHighEmphasis.withOpacity(0.2)
-              : _attachmentContainsFile &&
-                      messageInputController.attachments.isNotEmpty
-                  ? streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.2)
-                  : streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.5);
-        case 3:
-          return attachmentLimitCrossed
-              ? streamChatThemeData.colorTheme.textHighEmphasis.withOpacity(0.2)
-              : _attachmentContainsFile &&
-                      messageInputController.attachments.isNotEmpty
-                  ? streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.2)
-                  : streamChatThemeData.colorTheme.textHighEmphasis
-                      .withOpacity(0.5);
-        default:
-          return Colors.black;
-      }
-    }
-
-    return AnimatedContainer(
-      duration: _openFilePickerSection
-          ? const Duration(milliseconds: 300)
-          : const Duration(),
-      curve: Curves.easeOut,
-      height: _openFilePickerSection ? _kMinMediaPickerSize : 0,
-      child: SingleChildScrollView(
-        child: SizedBox(
-          height: _kMinMediaPickerSize,
-          child: Material(
-            color: _streamChatTheme.colorTheme.inputBg,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Row(
-                  children: [
-                    IconButton(
-                      icon: StreamSvgIcon.pictures(
-                        color: _getIconColor(0),
-                      ),
-                      onPressed: _attachmentContainsFile &&
-                              messageInputController.attachments.isNotEmpty
-                          ? null
-                          : () {
-                              setState(() {
-                                _filePickerIndex = 0;
-                              });
-                            },
-                    ),
-                    IconButton(
-                      iconSize: 32,
-                      icon: StreamSvgIcon.files(
-                        color: _getIconColor(1),
-                      ),
-                      onPressed: !_attachmentContainsFile &&
-                              messageInputController.attachments.isNotEmpty
-                          ? null
-                          : () {
-                              pickFile(DefaultAttachmentTypes.file);
-                            },
-                    ),
-                    IconButton(
-                      icon: StreamSvgIcon.camera(
-                        color: _getIconColor(2),
-                      ),
-                      onPressed: attachmentLimitCrossed ||
-                              (_attachmentContainsFile &&
-                                  messageInputController.attachments.isNotEmpty)
-                          ? null
-                          : () {
-                              pickFile(
-                                DefaultAttachmentTypes.image,
-                                camera: true,
-                              );
-                            },
-                    ),
-                    IconButton(
-                      padding: const EdgeInsets.all(0),
-                      icon: StreamSvgIcon.record(
-                        color: _getIconColor(3),
-                      ),
-                      onPressed: attachmentLimitCrossed ||
-                              (_attachmentContainsFile &&
-                                  messageInputController.attachments.isNotEmpty)
-                          ? null
-                          : () {
-                              pickFile(
-                                DefaultAttachmentTypes.video,
-                                camera: true,
-                              );
-                            },
-                    ),
-                  ],
-                ),
-                DecoratedBox(
-                  decoration: BoxDecoration(
-                    color: _streamChatTheme.colorTheme.barsBg,
-                    borderRadius: const BorderRadius.only(
-                      topLeft: Radius.circular(16),
-                      topRight: Radius.circular(16),
-                    ),
-                  ),
-                  child: Center(
-                    child: Padding(
-                      padding: const EdgeInsets.all(8),
-                      child: Container(
-                        width: 40,
-                        height: 4,
-                        decoration: BoxDecoration(
-                          color: _streamChatTheme.colorTheme.inputBg,
-                          borderRadius: BorderRadius.circular(4),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-                if (_openFilePickerSection)
-                  Expanded(
-                    child: DecoratedBox(
-                      decoration: BoxDecoration(
-                        color: _streamChatTheme.colorTheme.barsBg,
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      child: _PickerWidget(
-                        filePickerIndex: _filePickerIndex,
-                        streamChatTheme: _streamChatTheme,
-                        containsFile: _attachmentContainsFile,
-                        selectedMedias: messageInputController.attachments
-                            .map((e) => e.id)
-                            .toList(),
-                        onAddMoreFilesClick: pickFile,
-                        onMediaSelected: (media) {
-                          if (messageInputController.attachments
-                              .any((e) => e.id == media.id)) {
-                            setState(() => messageInputController.attachments
-                                .removeWhere((e) => e.id == media.id));
-                          } else {
-                            _addAssetAttachment(media);
-                          }
-                        },
-                      ),
-                    ),
-                  ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  void _addAssetAttachment(AssetEntity medium) async {
-    final mediaFile = await medium.originFile.timeout(
-      const Duration(seconds: 5),
-      onTimeout: () => medium.originFile,
+    final picker = StreamAttachmentPicker(
+      messageInputController: messageInputController,
+      onFilePicked: pickFile,
+      isOpen: _openFilePickerSection,
+      pickerSize: _openFilePickerSection ? _kMinMediaPickerSize : 0,
+      attachmentLimit: widget.attachmentLimit,
+      onAttachmentLimitExceeded: widget.onAttachmentLimitExceed,
+      maxAttachmentSize: widget.maxAttachmentSize,
+      compressedVideoQuality: widget.compressedVideoQuality,
+      compressedVideoFrameRate: widget.compressedVideoFrameRate,
+      onChangeInputState: (val) {
+        setState(() {
+          _inputEnabled = val;
+        });
+      },
+      onError: _showErrorAlert,
     );
 
-    if (mediaFile == null) return;
-
-    var file = AttachmentFile(
-      path: mediaFile.path,
-      size: await mediaFile.length(),
-      bytes: mediaFile.readAsBytesSync(),
-    );
-
-    if (file.size! > widget.maxAttachmentSize) {
-      if (medium.type == AssetType.video && file.path != null) {
-        final mediaInfo = await (VideoService.compressVideo(
-          file.path!,
-          frameRate: widget.compressedVideoFrameRate,
-          quality: widget.compressedVideoQuality,
-        ) as FutureOr<MediaInfo>);
-
-        if (mediaInfo.filesize! > widget.maxAttachmentSize) {
-          _showErrorAlert(
-            context.translations.fileTooLargeAfterCompressionError(
-              widget.maxAttachmentSize / (1024 * 1024),
-            ),
-          );
-          return;
-        }
-        file = AttachmentFile(
-          name: file.name,
-          size: mediaInfo.filesize,
-          bytes: await mediaInfo.file?.readAsBytes(),
-          path: mediaInfo.path,
-        );
-      } else {
-        _showErrorAlert(context.translations.fileTooLargeError(
-          widget.maxAttachmentSize / (1024 * 1024),
-        ));
-        return;
-      }
-    }
-
-    setState(() {
-      final attachment = Attachment(
-        id: medium.id,
-        file: file,
-        type: medium.type == AssetType.image ? 'image' : 'video',
+    if (_openFilePickerSection && widget.attachmentsPickerBuilder != null) {
+      return widget.attachmentsPickerBuilder!(
+        context,
+        messageInputController,
+        picker,
       );
-      _addAttachments([attachment]);
-    });
+    }
+
+    return picker;
   }
 
   Widget _buildMentionsOverlayEntry() {
@@ -1736,53 +1545,12 @@ class MessageInputState extends State<MessageInput> {
     });
   }
 
-  Widget _buildIdleSendButton(BuildContext context) => Padding(
-        padding: const EdgeInsets.all(8),
-        child: StreamSvgIcon(
-          assetName: _getIdleSendIcon(),
-          color: _messageInputTheme.sendButtonIdleColor,
-        ),
-      );
-
-  Widget _buildSendButton(BuildContext context) => Padding(
-        padding: const EdgeInsets.all(8),
-        child: IconButton(
-          onPressed: sendMessage,
-          padding: const EdgeInsets.all(0),
-          splashRadius: 24,
-          constraints: const BoxConstraints.tightFor(
-            height: 24,
-            width: 24,
-          ),
-          icon: StreamSvgIcon(
-            assetName: _getSendIcon(),
-            color: _messageInputTheme.sendButtonColor,
-          ),
-        ),
-      );
-
-  String _getIdleSendIcon() {
-    if (_commandEnabled) {
-      return 'Icon_search.svg';
-    } else {
-      return 'Icon_circle_right.svg';
-    }
-  }
-
-  String _getSendIcon() {
-    if (widget.editMessage != null) {
-      return 'Icon_circle_up.svg';
-    } else if (_commandEnabled) {
-      return 'Icon_search.svg';
-    } else {
-      return 'Icon_circle_up.svg';
-    }
-  }
-
   /// Sends the current message
   Future<void> sendMessage() async {
     var text = messageInputController.text.trim();
-    if (text.isEmpty && messageInputController.attachments.isEmpty) {
+    final attachments = messageInputController.attachments;
+
+    if (text.isEmpty && attachments.isEmpty) {
       return;
     }
 
@@ -1805,7 +1573,7 @@ class MessageInputState extends State<MessageInput> {
     if (widget.editMessage != null) {
       message = widget.editMessage!.copyWith(
         text: text,
-        attachments: messageInputController.attachments,
+        attachments: attachments,
         mentionedUsers: messageInputController.mentionedUsers
             .where((u) => text.contains('@${u.name}'))
             .toList(),
@@ -1814,7 +1582,7 @@ class MessageInputState extends State<MessageInput> {
       message = (widget.initialMessage ?? Message()).copyWith(
         parentId: widget.parentMessage?.id,
         text: text,
-        attachments: messageInputController.attachments,
+        attachments: attachments,
         mentionedUsers: messageInputController.mentionedUsers
             .where((u) => text.contains('@${u.name}'))
             .toList(),
@@ -1960,7 +1728,7 @@ class MessageInputState extends State<MessageInput> {
   void didChangeDependencies() {
     _streamChatTheme = StreamChatTheme.of(context);
     _messageInputTheme = MessageInputTheme.of(context);
-    if (widget.editMessage == null) _startSlowMode();
+    if (widget.editMessage == null && _timeOut <= 0) _startSlowMode();
 
     if ((widget.editMessage != null || widget.initialMessage != null) &&
         !_initialized) {
@@ -1969,141 +1737,4 @@ class MessageInputState extends State<MessageInput> {
     }
     super.didChangeDependencies();
   }
-}
-
-class _PickerWidget extends StatefulWidget {
-  const _PickerWidget({
-    Key? key,
-    required this.filePickerIndex,
-    required this.containsFile,
-    required this.selectedMedias,
-    required this.onAddMoreFilesClick,
-    required this.onMediaSelected,
-    required this.streamChatTheme,
-  }) : super(key: key);
-
-  final int filePickerIndex;
-  final bool containsFile;
-  final List<String> selectedMedias;
-  final void Function(DefaultAttachmentTypes) onAddMoreFilesClick;
-  final void Function(AssetEntity) onMediaSelected;
-  final StreamChatThemeData streamChatTheme;
-
-  @override
-  _PickerWidgetState createState() => _PickerWidgetState();
-}
-
-class _PickerWidgetState extends State<_PickerWidget> {
-  Future<bool>? requestPermission;
-
-  @override
-  void initState() {
-    super.initState();
-    requestPermission = PhotoManager.requestPermission();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (widget.filePickerIndex != 0) {
-      return const Offstage();
-    }
-    return FutureBuilder<bool>(
-      future: requestPermission,
-      builder: (context, snapshot) {
-        if (!snapshot.hasData) {
-          return const Offstage();
-        }
-
-        if (snapshot.data!) {
-          if (widget.containsFile) {
-            return GestureDetector(
-              onTap: () {
-                widget.onAddMoreFilesClick(DefaultAttachmentTypes.file);
-              },
-              child: Container(
-                constraints: const BoxConstraints.expand(),
-                color: widget.streamChatTheme.colorTheme.inputBg,
-                alignment: Alignment.center,
-                child: Text(
-                  context.translations.addMoreFilesLabel,
-                  style: TextStyle(
-                    color: widget.streamChatTheme.colorTheme.accentPrimary,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            );
-          }
-          return MediaListView(
-            selectedIds: widget.selectedMedias,
-            onSelect: widget.onMediaSelected,
-          );
-        }
-
-        return InkWell(
-          onTap: () async {
-            PhotoManager.openSetting();
-          },
-          child: Container(
-            color: widget.streamChatTheme.colorTheme.inputBg,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                SvgPicture.asset(
-                  'svgs/icon_picture_empty_state.svg',
-                  package: 'stream_chat_flutter',
-                  height: 140,
-                  color: widget.streamChatTheme.colorTheme.disabled,
-                ),
-                Text(
-                  context.translations.enablePhotoAndVideoAccessMessage,
-                  style: widget.streamChatTheme.textTheme.body.copyWith(
-                    color: widget.streamChatTheme.colorTheme.textLowEmphasis,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-                const SizedBox(height: 6),
-                Center(
-                  child: Text(
-                    context.translations.allowGalleryAccessMessage,
-                    style: widget.streamChatTheme.textTheme.bodyBold.copyWith(
-                      color: widget.streamChatTheme.colorTheme.accentPrimary,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class _CountdownButton extends StatelessWidget {
-  const _CountdownButton({
-    Key? key,
-    required this.count,
-  }) : super(key: key);
-
-  final int count;
-
-  @override
-  Widget build(BuildContext context) => Padding(
-        padding: const EdgeInsets.all(8),
-        child: DecoratedBox(
-          decoration: BoxDecoration(
-            color: StreamChatTheme.of(context).colorTheme.disabled,
-            shape: BoxShape.circle,
-          ),
-          child: SizedBox(
-            height: 24,
-            width: 24,
-            child: Center(
-              child: Text('$count'),
-            ),
-          ),
-        ),
-      );
 }