+++ packages/stream_chat_flutter_core/test/channels_bloc_test.dart
@@ -10,7 +10,7 @@ import 'mocks.dart';
 
 void main() {
   setUpAll(() {
-    registerFallbackValue<PaginationParams>(const PaginationParams());
+    registerFallbackValue(const PaginationParams());
   });
 
   List<Channel> _generateChannels(