+++ packages/stream_chat_persistence/lib/src/dao/member_dao.g.dart
@@ -6,7 +6,7 @@ part of 'member_dao.dart';
 // DaoGenerator
 // **************************************************************************
 
-mixin _$MemberDaoMixin on DatabaseAccessor<MoorChatDatabase> {
+mixin _$MemberDaoMixin on DatabaseAccessor<DriftChatDatabase> {
   $MembersTable get members => attachedDatabase.members;
   $UsersTable get users => attachedDatabase.users;
 }