+++ packages/stream_chat_persistence/lib/src/dao/pinned_message_dao.g.dart
@@ -6,7 +6,7 @@ part of 'pinned_message_dao.dart';
 // DaoGenerator
 // **************************************************************************
 
-mixin _$PinnedMessageDaoMixin on DatabaseAccessor<MoorChatDatabase> {
+mixin _$PinnedMessageDaoMixin on DatabaseAccessor<DriftChatDatabase> {
   $PinnedMessagesTable get pinnedMessages => attachedDatabase.pinnedMessages;
   $UsersTable get users => attachedDatabase.users;
 }