+++ packages/stream_chat_persistence/lib/src/entity/reads.dart
@@ -1,5 +1,5 @@
 // coverage:ignore-file
-import 'package:moor/moor.dart';
+import 'package:drift/drift.dart';
 
 /// Represents a [Reads] table in [MoorChatDatabase].
 @DataClassName('ReadEntity')