+++ packages/stream_chat_persistence/test/src/dao/channel_dao_test.dart
@@ -1,13 +1,13 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_persistence/src/dao/channel_dao.dart';
-import 'package:stream_chat_persistence/src/db/moor_chat_database.dart';
+import 'package:stream_chat_persistence/src/db/drift_chat_database.dart';
 
 import '../../stream_chat_persistence_client_test.dart';
 
 void main() {
   late ChannelDao channelDao;
-  late MoorChatDatabase database;
+  late DriftChatDatabase database;
 
   setUp(() {
     database = testDatabaseProvider('testUserId');